Fix #6932 - login isn't working
