+++ frontend/src/network/genericJSONRequest.js
@@ -7,7 +7,7 @@ import { API_URL } from '../config';
  * @param {RequestInit} [init={}}] Any specific init options you want to pass the fetch (such as an {@link AbortSignal})
  * @returns {Promise<*>} A promise that returns a JSON or an error
  */
-export function fetchExternalJSONAPI(url, init = {}): Promise<*> {
+export function fetchExternalJSONAPI(url, init = {}) {
   if (!init.headers) {
     init.headers = { 'Content-Type': 'application/json' };
   }
@@ -23,7 +23,7 @@ export function fetchExternalJSONAPI(url, init = {}): Promise<*> {
     });
 }
 
-export function fetchLocalJSONAPI(endpoint, token, method = 'GET', language = 'en'): Promise<*> {
+export function fetchLocalJSONAPI(endpoint, token, method = 'GET', language = 'en') {
   const url = new URL(endpoint, API_URL);
   let headers = {
     'Content-Type': 'application/json',
@@ -74,7 +74,7 @@ export function pushToLocalJSONAPI(
   token,
   method = 'POST',
   language = 'en',
-): Promise<*> {
+) {
   const url = new URL(endpoint, API_URL);
   return fetch(url, {
     method: method,
